+++ tools/ncnnoptimize.cpp
@@ -94,7 +94,7 @@ class DataReaderFromEmpty : public ncnn::DataReader
 {
 public:
     virtual int scan(const char* format, void* p) const { return 0; }
-    virtual int read(void* /*buf*/, int size) const { return size; }
+    virtual size_t read(void* /*buf*/, size_t size) const { return size; }
 };
 
 class NetOptimize : public ncnn::Net
